Actividad2

+++ UD4-Modelo-Vista-Controlador (MVC)/Actividades/Actividad2/Actividad1/Models/Pokemon.cs
@@ -7,7 +7,7 @@ namespace Actividad1.Models
     {
         public int PokemonId { get; set; }
         public string nombre { get; set; }
-        public float peso { get; set; }
-        public float altura { get; set; }
+        public double peso { get; set; }
+        public double altura { get; set; }
     }
 }
